MB-63908: Add support for locking the builtin admin user
The builtin admin user can be locked with POST /controller/lockAdmin,
provided the request is made using the localtoken auth, and unlocked
equivalently with POST /controller/unlockAdmin.

This protects against unintentionally losing all access to the cluster,
since in order to lock the admin, you must have access to the actual
server, and so you'll also still be able to unlock the admin.

Change-Id: I0bef6c7098406843f7d27a9f53cba707867b516e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/218197
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>