MB-59113: Alerts for data disk usage guardrail
When the data disk limit is configured, such that mutations are not
permitted once a disk reaches that limit, the administrator needs to be
alerted as this limit is approached.

They will initially get a warning alert, once a node is within 10% of
the disk limit. Once the limit is reached, they will get a critical
alert.

If the data disk limit is not configured, we will still alert the
admin once a disk reaches 75% (configurable with maxDataDiskUsedPerc in
the alert limits endpoint).

Change-Id: Ie2b09b293d0933a87d3dfffcc3a66d993ff55a36
Reviewed-on: https://review.couchbase.org/c/ns_server/+/217786
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>