MB-61292: Add support for NULL dek
KV stat encryption-key-ids can return special key id = "unencrypted"
This means that some data in that vbucket is unencrypted.

It is also possible to use "unecrypted" as a key to drop in
the compact_vbucket command. In this case it means that vbucket
must encrypt all the data that is still unencrypted.

In ns_server that special key is translated to a key with special
name: NULL_DEK_ID.

Change-Id: I7198d883951d8252b50b59f4997ec9b73aaccb1b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/217827
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>