MB-63869: Fix auto_failover crash when service safety check fails
If the service safety check fails (either because the node cannot be
safely failed over or because the service does not respond to the
safety check) and we have NO safe nodes to fail over then we currently
crash the auto_failover process as we attempt to use the
{ok, UnsafeNodes} result of try_autofailover/5 as a replacement for the
gen_server state. This results in a bad record crash. We need to return
the state here and ensure that we call log_unsafe_node/2 to report the
error.

Change-Id: I040ae701459b372debee7b5bd81188b1a727080a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/217950
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Peter Searby <peter.searby@couchbase.com>