MB-64005: Restrict disk checks to nodes with data service
The original guardrail was overly restrictive, in the aim of doing
whatever we can to mitigate potential issues caused by disk space
running out. This meant that when any node’s data disk hit the
restriction threshold, all nodes would stop permitting mutations in the
data service. The idea was that data service mutations can contribute
to disk issues for other services, even if the disks are different,
because the other service may ingest mutations from the data service.

However, this makes the behaviour potentially less intuitive, and has a
significant risk of false positives causing loss of data service
availability (such as CBSE-17251). We initially decided to keep the
behaviour the same, since it at least provided additional protection,
even if it was somewhat overkill.

Now that the guardrail is being enabled for self-managed, this decision
was re-evaluated, and based on discussion with PM, we decided that the
more intuitive behaviour is to only make the check on data service
nodes.

In future we may wish to improve things with service-specific logic,
which could then mitigate the issue that this behaviour was originally
intended for.

Change-Id: I3876675904123cadd8f5739cfba08dde0e5ce894
Reviewed-on: https://review.couchbase.org/c/ns_server/+/218533
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>