MB-64034: Don't assert that ns_bucket:wait_for_bucket_shutdown is ok
This causes a badmatch and "Unexpected server error" on bucket deletion
if the deletion times out. We handled this differently prior to 7.6, we
would respond with a 500 code but an additional context of "Bucket
deletion did not yet complete".

Remove the ok assertions and return the old value back to the caller.

Change-Id: I5d42c63cb6cc02354bda93a295c0cbecc1faceaa
Reviewed-on: https://review.couchbase.org/c/ns_server/+/218746
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>