MB-61292: [dets fork] Copy dets files
This commit copies the dets files to cb_dets files. The files are
copied as is without any changes. The files are copied to a new
directory cb_dets in the apps directory.

The main purpose of copying the files is to modify them in the
subsequent commits to support encrypted dets files.

Vanilla dets does not support encryption. The changes in the
subsequent commits will add support for encrypted dets files.

cb_dets will use configDek to encrypt and decrypt the dets files.

Note that we can't simply encrypt data before writing to dets.
If we do so, we won't be able to use any lookups, or selects.

Change-Id: I3f2ba3ccb1f33384ec6266696acc269f1c0edb4d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/218971
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>