Fix menelaus_web_settings2 subtypes formatter search
When searching for a subtype we should check all availvable
type_spec functions.

For example, say we have two type_spec functions: [F1, F2],
and we have a complex type Type that depends on another type SubType.
Let's assume that Type is defined in F2, while Subtype is defined
by F1.
Old behavior: when we have found definition for Type, we will
not check F1 for definition of Subtype, which is a mistake.
New behavior: reset typs spec funs to be used for search when
we start searching for Subtype definition.

Change-Id: I897c603a18bb05e22e23082d6b3ec7b64af17feb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219081
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>