MB-61292: Log GC accounting for stopped ale disk sinks
Although we only stop/start ale disk sinks on server start, it is
possible in code to stop/start sinks at any time. The purpose
of this change is to account for the used deks of stopped disk
sinks, if in use query comes in while any sink is in stopped
state. This is done by saving the used deks of the disk sink
prior to stopping it.

Change-Id: I4ade4c4fe4d2b175d32dca5936984d67bde2a8ab
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219238
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>