cluster_tests: Don't wait_for_balanced when unnecessary
The wait_for_rebalance function by default was waiting for the cluster
to be balanced, however there are a lot of cases where it wasn't
actually needed, and was unnecessarily adding a substantial amount
of test prep time.

To avoid wasting resources by default, we should default to not waiting
for the cluster to be balanced, and only opt-in to the wait when
applicable.
In addition, we can skip the wait when the rebalance gives an error.

To make it clear when this wait is/isn't being performed, explicit
opt-out as well as opt-in is added in each usage. This makes it obvious
what the code does without digging into the cluster_run_lib function.

For now, the only cases where we will wait for the cluster to be
balanced are those where we have just performed a rebalance,
specifically in Cluster.rebalance(), ServicesTopologyTests.rebalance()
and when connecting a cluster in cluster_run_lib.

The other cases are generally where we care only about waiting for
the rebalance to complete as we are about to do something that can't
be performed during a rebalance, such as bucket crud operations.

Change-Id: Id70975751cf72bac10e187f7ccb467add818b918
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219319
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>