MB-61292: Fix decrypt functions
the case when magic is incomplete and when header is incomplete
are different:
When we have full magic we know that the file is encrypted,
so if other part of the header is missing it is invalid file.
When magic is incomplete we don't know if the file is encrypted
or not.

Change-Id: I407e62adda7deaca634078819ff9cfc5b3c4a697
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219294
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>