MB-63417: New invalidHlcStrategy bucket setting
This adds the invalidHlcStrategy bucket setting with possible values
as "error", "ignore", "replace". The default value for existing buckets
e.g. before upgrade is "ignore" which is the same as ns_server not
specifying the setting at all in the bucket config. The default value
for new buckets is "error".

Change-Id: I5699d7af16f681f43a48a417ef3b31f8169cc60d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219331
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>