MB-31823: Use json errors structure for reused password
By using the `{"errors": {"password": ...` structure, we are not only
keeping consistent with other errors given when changing the password,
but it always ensures that the UI behaves correctly.

Without this change, the UI hangs after receiving the error. With the
change it correctly shows the error message.

Change-Id: Iafdea8dc4ca52cefa4c52a3c6f3d848c11583c49
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219369
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>