MB-62214: Bucket Setting Cross Cluster Versioning.
Adds a new property in Bucket Settings dialog: Enable Cross Cluster Versioning.
The property is also displayed in bucket details section from the
buckets list.
It is not visible when creating a bucket because
it cannot be set while the bucket is created -
it is visible only for editing a bucket.
After the property is enabled for a bucket it cannot be disabled back
for that bucket.

Change-Id: I005a4fd6546bc680d4c8cff9779546e36e45fbe5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219444
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>