MB-59704: Rename memcached -> kv  ...
... to make consistent with how it's referred to elsewhere in the code.
This only changes how we refer to the port/service to make them
consistent. Before, it seems like we used both names, depending on the
situation or when it was introduced. The binary used is still
"memcached" and so no other tools need to be modified for this change.
There are also no rpc calls across nodes that use these labels so we
shouldn't have any cluster compatibility issues.

Change-Id: I31ab78596542c734233c74c9d27398ce5e7d723f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219456
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>