MB-59704: Main cgroup impl
This is the main implementation for the cgroups v2 integration for
provisioned. The main design is relatively simple and involves spawning
processes directly into the cgroups where they will run, or are moved
after being spawned. The config profile is used to determine some of the
"knobs" that can be used to adjust things.

* Profile / Settings:
For example here is the cgroup_base_path which is the base directory
where all the services will be placed:
```
{cgroup_base_path,
    "/sys/fs/cgroup/system.slice/couchbase-server.service/services"}.
```

Then there are the lists of cgroups_hard and cgroups_soft which
determine which services will have hard/soft limits applied. The value
applied is the memoryQuota unless an OVERRIDE has been set for the
service:
```
{cgroups_memory_hard, [n1ql, fts]}.
{cgroups_memory_soft, [n1ql, fts, prometheus, kv, projector,
                       ns_server, backup, index, cbas, goxdcr,
                       eventing, cont_backup]}.
```

* Overrides:
Overrides are used to give a different numeric value that aren't tied to
the memoryQuota of the service. It can also be used to give limits to
services that do not have the concept of a memoryQuota like ns_server
and promtheus. Overrides are meant to be node-local.

Longer-term changes to which service to place in hard/soft group can and
should be done through the profile and not the overrides.

* Goport:
Goport was modified to accept an argument "cgroup" which is a string of
a directory path that represents the cgroup to place the process in. The
cgroup path is then passed to the code that launches the process and
modifies the SysProcAttr to include the file descriptor of the opened
directory path as well as the UseCgroupFD to true. This modification
only affects linux since it's a platform dependent feature.

The goport processes are all spawned directly into the cgroup while
ns_server, kv are both launched using a normal erlang port and then
immediately moved to the correct cgroup. This is somewhat non-ideal
however there is no easy way to do it the goport way while still using
standard ports. It would have involved creating a shim or similar and
basically re-creating many aspects of goport, which seems like a waste
of time (for limited return). To further investigate whether all ports
can be ran using goport (to simplify this and all port handling) you can
see MB-64818 which will track that task.

* Manager:
The main process managing this in ns_server is the ns_cgroups_manager
gen_server that checks the current state of the cgroups and brings them
into compliance with that state. This means it will move-back processes
that are moved out of their desired cgroup. More importantly, it will
check the actual values set to either memory.high/memory.max and keeps
them in line with what we think they should be in ns_server.

* Caveats:
This is a feature for linux only, with provisioned profile. It also
requires cgroups v2 on the system and not using the combined v1/v2
setup. Cgroups v1 has been deprecated and is slowly being removed from
all distros. It's a bit of a slow process especially for LTS (long term
support) releases.

Change-Id: I9342219411ecad9cfa7b01a46aae0cfe9a2e4359
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219457
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>