MB-64005: Fix index disk usage guardrail logic
The disk usage guardrail for the index service has some of its logic
already implemented, even though indexing hasn't actually implemented
its side of the logic for any guardrails. Because of this discrepancy,
we disabled the configuration for index guardrails in this previous
commit: 2e8797001caeb6aeeb2ddc8ce0b705310a7a11e6.

However, as part of this change, we moved the configuration for the
index disk usage guardrail to a separate dedicated key within the
index guardrails sub-config, without making a corresponding change
in the logic of the regular checks.
This wasn't spotted, because the tests used the old config layout, and
it all just worked out anyway.

While it would be easier to just delete all the partially complete
indexing guardrails code, this would make it more of a nuisance to add
back later, so instead we are trying to keep the code and tests in
place.

To resolve this, we need to refactor the logic substantially, so that
we can reuse most of the existing disk usage check code, while
splitting it out for a separate index service check.

Change-Id: Id6341e70894e7630f6914a8933b394446eb3df99
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219504
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>