MB-64235: XDCR mobile setting.
Adds a new replication setting: 'mobile'.
Valid API values are 'Active' and 'Off', we are showing the values
in a toggle, so we handle them as `true`/`false`.
The setting is visible on Add/Edit Replication pages.
The toggle is disabled when the source bucket has the property
crossClusterVersioning set to false.

Change-Id: I687d14f479332225683227b17331b79514aa967e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219575
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222020
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>