MB-61292: Validate for complete encrypted chunks
Anytime the logger considers an encrypted file for continuation upon
start, we validate the file for the presence of complete encrypted
chunks. If validation fails, file is rotated and new encrypted file
is started for logging. This ensures that we don't continue writing
encrypted chunks to a file that may have partial chunks in it.

Change-Id: Ia869beccfff14684fbfcc99059165c321dfded87
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219968
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>