MB-64365: Monitor process instead of using bucket_events
Using buckets_events in this way is racey. If we want the ns_memcached
process to be terminated, we need to wait for that, not for the
buckets_events notification.

This racing can cause a rebalance to wait 60s at the start before doing
anything, which is poor behaviour in production, and a nuisance for our
tests (e.g. MB-64287)

Change-Id: I75291bd0e658327fa1fa8d99a3f734335aa9abc3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/219959
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>