MB-64587: Add app_telemetry_pool for websocket connections
This module provides an interface for handling multiple app telemetry
websocket connections.

It provides handle_connect/2 for taking a mochiweb request, upgrading
it to a websocket connection, and handling data sent and received via
this websocket.

The pool has a maximum number of connections, after which it will not
accept further connections.

The pool will automatically respond to ping messages sent by the client
with a pong message.

The pool will drop the websocket connection if it receives a frame that
it doesn't have a handler for.
A handler is added with call/3, specifying the connection by its pid,
alongside a binary body to send to the client, and a timeout to expect
a response within. Once a response is received from the client, the
call will return with the payload received.

The pool also permits manually dropping a connection, and getting the
list of pids associated with connection in the pool.

Change-Id: Ie93efead67be6285fbf86d617419243504058710
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220402
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>