MB-62381: Allow timeout in test
Rather than comment out the ns_rebalance_observer:rebalance test which
was done previously to handle occasional timeouts we'll wrap the call
in a try/catch and allow the timeout.

This same may occur for ns_rebalance_observer:failover but I'd like to
see if that is true before also wrapping that test in a try/catch.

Change-Id: I6f330040254b7bb850b22b74de165de58c59ecc4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220354
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>