MB-64481: Ensure KV is notified when user is locked via PATCH
When a user is locked via PUT, we update both the auth_version and the
user_version. This triggers the user to be removed from the
memcached.rbac file, but not the isasl.pw file. However, memcached only
needs the user removed from one of the files, in order to consider the
user to not exist, so this worked fine.

The issue is that for PATCH updates we correctly only update the
auth_version, but then this means that the memcached.rbac file still
contains the user, so memcached allows the user to still authenticate.

We should remove the locked user from both files, so that even if only
the auth_version changes, we still immediately prevent the user from
continuing to authenticate.

This commit also adds tests to cover users being locked from KV's
perspective. We do this via the python SDK, so that dependency has
been added to cluster_tests/requirements.txt. The dependency has been
specified without a version, since unlike other libraries, we probably
do care if the latest version breaks, so we ideally want to test with
the latest version.

Change-Id: I7ae54b11e0ddf77e44acd832054b023cd0b09a36
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220394
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>