MB-62381 Clean up ns_rebalance_observer test changes
Now that the failures of the ns_rebalance_observer tests are under
control I'm removing some of the interim changes used to triage the
issue. The try/catch of the timeout is left as a "canary in the
coalmine" to detect any future occurrences.

Change-Id: Iccf5b547e933c0622da449bef57a4ad4522171a8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220563
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>