MB-64507, MB-61292: [cluster tests] Increase poll...
... interval for native encryption tests

Currently we wait for 10 seconds exactly, but the problem is that
cb_cluster_secrets also has a 10 second retry interval, so it is
possible that poll finishes right before retry timer fires.
In some tricky cases, it is ok that logic like DEK reencryption
requires retrying. One of such scenarios is the following:
1. Secret gets rotated (so there is a new KEK, and DEKS need to be
   reencrypted);
2. In order to reencrypt DEKs, 2 jobs need to be run:
    a. Store new KEK on disk;
    b. Reencrypt DEKs with the new KEK
   So these jobs are about to get scheduled.
3. It turns out that job (b) was already scheduled before,
   so after scheduling of (a), it ends up in the queue after (b).
4. So (b) starts before (a) and fails because the new KEK is not
   yet stored on disk.
5. (a) finishes successfully, but (b) will be retried after 10
   seconds (retry interval).

DEK reencryption gets scheduled, but there is a

Change-Id: Ic033e2eed93dac9cf078dff417af742f418be350
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220622
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>