MB-61292: Add integrity check for stored keys
This commit adds an integrity check for stored keys. The check is
performed every time the stored keys are loaded from disk.
The check is done by comparing the stored key proof with the proof
that is calculated using token from the tokens file.
Proof for Key with KeyId is calculated the following way:

 Proof := TokenUUID + ENCRYPT(SHA512(Token + KeyId) with Key)

Token file contains all the tokens that are considered trusted for
the integrity check. Tokens are stored in the following format:

 TokenUUID1:Token1\n
 TokenUUID2:Token2\n
 ...

Token file is encrypted with current config dek.

If configuration encryption is disabled, tokens are stored in plain,
so there basically no integrity protection for keys in this case.

When configuration encryption gets enabled, we generate new
configuration key first, then we generate new token, and then we
store that new token in the tokens file encrypted with the newly
generated congifuration key. By doing so we are making sure that
new token is never stored in plain text.
Old tokens (that were stored in plain text) are removed from the
tokens file once all the existing keys have their proofs recalculated
with the latest token.

Rotation of tokens is performed on any config key generation, or
config encryption enabling/disabling.

Change-Id: I3de47705964337f20b4eb035115210549979f5c5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220792
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>