MB-64587: Advertise appTelemetryPath
When app telemetry is enabled, this is advertised to clients by
including appTelemetryPath in each node's entry in nodesExt. When a
node is not accepting connections, this fields will not be included.
Currently each node will accept connections precisely when app
telemetry is enabled, but in future we may decide to vary it on a per
node basis, for instance based on the number of connections used up.

'appTelemetryPath' specifies to the clients the relative path to
establish a websocket connection on that node.

Change-Id: I13fb39abc9fa505432782039f696ae6e7846462c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221738
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>