MB-64595: Correctly format trackedGroups in userActivity
The default formatter was representing the group names as lists of
integers, rather than strings. We should format the groups as an array
of strings.

Also improves the JSON parsing to expect an array of strings, rather
than a comma-separated string.

The tests have been updated to use the JSON format, so we can simply
check if the activity settings were configured as intended.

Change-Id: I8ed06f667f00adcff0d4cd734b695ba2b48020f7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220833
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>