MB-64481: Add locked users back to memcached.rbac
We only need to delete them from one out of memcached.rbac and
isasl.pw, and since we don't always delete them from memcached.rbac
(such as in a PATCH request, since the user_version doesn't change), it
is better to be consistent and just never delete them from
memcached.rbac.

While it could be argued that the most consistent thing to do would be
to remove the user from both files, this would mean that the
user_version would need to change on the PATCH request, which makes
things more complicated, as we would then need to update multiple
versions for the same user event.

This commit also adds a comment clarifying how locked users work for
kv, alongside the check in memcached_passwords.erl.

Change-Id: Ibabf77e5053e6835b1413fb6e4c2a5eaaaf98725
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220838
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>