MB-61292: [gosecrets] Reset list of touched keys when...
... switching to the next key in the list when reencrypting keys.

We use tochKeys to keep track of keys that have been touched during
decryption of a key. This is used to detect cycles on the key chain
and prevent infinite loops.
When we switching to the next key, the list of touched keys should
be reset to empty.
Otherwise it will detect a cycle in the following configuration:
key1 encrypted by key3
key2 encrypted by key3

Change-Id: Ie7f1e13c48edec01516ba74dd02ab91aa55fa938
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220902
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>