MB-62381: decrease recovery test output
Prior to this change the recovery:compute_recovery_map_test would
generate > 1 million lines of logging. This would, on certain machines
(e.g. jenkins servers) overwhelm the logging infrastructure and lead to
timeouts in downsteam tests using ?log_debug

The number of lines of output for the test decrease from > 1 million to
about 5,600.

In order to get verbose output for the test, change 'true' to 'false'
in recovery.erl in this code:

     Map = mb_map:generate_map(EmptyMap, NumCopies - 1, Servers,
                               [{terse_output, true}]),

Change-Id: Ib0b960a7964546a69e872d1a29713747774180b6
Reviewed-on: https://review.couchbase.org/c/ns_server/+/220889
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>