MB-61292: [gosecrets] Add read-only mode for gosecrets
so it doesn't create any keys or tokens on disk when it is
started for read-only purposes

The biggest concern here is the following scenario:
Two parallel gosecrets instances can race on creating keys
on disk. In this case on of the instances can end up having
one key in memory, and another key on disk, which is catastrophic.

By adding read-only mode we can avoid this situation, because only
ns_server will start gosecrets in read-write mode, so all other
instances will be read-only and will not create any keys.

Change-Id: I7ce935622ab8ca1cade6d6d66df79bcf7d08a896
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221039
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>