MB-61292: Move logs and audit deks to config dir
It is always normal to remove logs and audit folder. It is very
easy to remove the deks folder with them, and ideally it should also
be ok, but is not that easy to handle key removal properly in this
case (we don't know if it is a hard error that we can't read the key
or it is ok, and we should ignore it). It seems to be much easier to
move deks to config dir and assume that all deks should always be
present instead of assumming that some deks can suddenly disappear
at any moment

Change-Id: I9232038202e167284c344e4dbf8a2b9e5e227c09
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221169
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>