MB-61292: Make sure we can read all deks at startup
Before this commit we ignored read key errors, because we
had to support the case when log dir is removed together
with log deks. Now since log deks are stored in config dir,
they can't be removed when logs are removed, so it should
be save to assume that deks must be always readable.

There is another scenario that needs to be kept in mind:
say we have a dek encrypted by aws key, and that aws key
is unavailable at startup, so we can't read that dek.
There are two ways to handle that:
1. Continue to start up, but retry reading deks later;
2. Fail to start up.

Option #1 is hard to implement as the code that uses that dek
should handle the case when dek is not available.

This is another reason why this commit implements option #2.
Note that this scenario was not supported before this commit.

Change-Id: Ib01c009957ae7f413428b38c6f2c32bb19f193db
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221170
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>