MB-64082 - UI support for RBAC role changes
There are three new RBAC roles:

- Local User Admin - may manage local users, but may not view any of
  the other tabs of the Security section of the UI.

- External User Admin - may manage LDAP/SAML users, but may not view
  any other tabs of the Security section of the UI.

- Security Admin - The security admin is not permitted to manage
  users, however the security admin may configure LDAP and SAML.

Of note, with Morpheus the security role admin.security.external is
being removed, in favor of admin.users.external.

This required changes to the tabs in the Security section of the UI:

- The Users & Groups tab is now hidden unless the user has one of the
  User Admin roles.

- The rest of the Security tabs are now hidden unless the user has the
  Security Admin role.

- LDAP configuration was previously in a dialog box accessibly from
  the Users & Groups tab, however since the roles have been split,
  this configuration has been moved to its own tab.

Testing	was done interactively,	creating users with each of the	roles
and testing the visibility and functionality. There are still problems
with REST endpoint permissions, which are being fixed as part of an
ns_server issue.

Change-Id: Id5f61f3a6b0d7c30642259face8203044d511e33
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221243
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>