MB-64598: Delete user activity from ets when user deleted
When a user is deleted, it will leave behind the activity timestamp in
the activity_tracker ets table on each node. This means that if a new
user is added with the same name, it would likely revive the old
timestamp, which was for the previous user.

To avoid having to call out to each node, telling them to delete any
activity timestamp, when deleting a user, instead it is simpler to
have the gen_server subscribe to user_storage_events, and clear up
unwanted timestamps on each user_version change.

This fix means we can remove the manual cleanup of timestamps from the
activity tracking cluster tests.

Change-Id: Ie4317a1787167d8d4a33cd23c713098c05c7706f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221434
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>