MB-64163: Encrypt Logs in Collect Logs page.
Adds a new option to encrypt the logs in Collect Information page.
A password must be provided to encrypt the logs.
A warning is visible under the checkbox when
encryption at REST for logs is enabled but the
newly added checkbox is unchecked, since in this case
the generated logs are not encrypted.

Change-Id: I3f4d564f53de23e5d6dd55b10ca3d4d3903b4a1c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221523
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>