MB-64754: new /settings/rbac API
The /setting/rbac endpoint is provide so the UI doesn't have to call
"security" related APIs to determine if ldap, saml, and saslauthd
are enabled when the user is a local/external user admin.

$ xcurl localhost:9000/settings/rbac | jq
{
  "ldapEnabled": false,
  "samlEnabled": false,
  "saslauthdAuthEnabled": false
}

The UI currently calls /settings/ldap, /settings/saml,
/settings/saslauthdAuth to get the info. But these APIs will have
their permission changed in a subsequent patch to disallow the user
admins from using them as they return "sensitive" security info.

Change-Id: Iabcf98b255788f336c7e1e2100552d06eaf0e392
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221716
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>