MB-59704: cgroupv2 tests
Tests for cgroup v2 that do not require cgroups v2 or an otherwise
compatible system. There is no dependency on operating system or
configuration. Ideally, it would be nice to have functional tests from
the point of view of cluster-tests, but we will need to wait until the
CI machines can support it.

Change-Id: I1812bc41c364d74b73e168e863c96a1ea42496a6
Reviewed-on: https://review.couchbase.org/c/ns_server/+/221947
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>