MB-50503: Fix serverTime timezone in /logs
We currently report the serverTime with 'Z', implying that it is UTC,
when in fact it is in the server timezone.

It should still be given in server timezone, as the name is
'serverTime', but the actual timezone should be specified.

This is gated behind a cluster compat mode change, to maintain
consistent reporting of times across the cluster.

Change-Id: I21e683291f9bf4c9a0a0f766f3fbd20c8fa5c410
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222312
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>