MB-50503: Fix serverTime timezone for alerts in /pools/default
We currently report the serverTime with 'Z', implying that it is UTC,
when in fact it is in the server timezone.

It should still be given in server timezone, as the name is
'serverTime', but the actual timezone should be specified.

We gate this behind a cluster compat mode change, not because it is
necessary to avoid bugs, but just to maintain consistent reporting of
times across the cluster. Without the cluster compat check, we could
have old nodes nodes reporting Z (UTC), with new nodes reporting the
actual timezone, which will likely cause confusion, since the times
will be entirely inconsistent.

Change-Id: I10a647c0d57c6902ccf4dd463ceef09fe6c9714a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222310
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>