MB-64776: Fix wrong timezones in user details
We currently report password_change_timestamp and last_activity_time
with 'Z', implying that they are UTC, when in fact they are in the
server timezone.

They should still be given in server timezone for consistency, but the
actual timezone should be specified.

This is gated behind a cluster compat mode change, to maintain
consistent reporting of times across the cluster.

Change-Id: Ia92c7be34123f7bad01cfb2d9035b5bb3a23a10d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222311
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>