MB-64587: Disconnect connections when the max is decreased
If an administrator wishes to decrease the maximum number of app
telemetry connections per node, it might be because they are facing
issues with that number of connections. We should immediately close
excess connections when they do this, in the hope of helping with the
issue.

This functionality will be used in a later commit to be able to
disconnect all connections when app telemetry is disabled.

Also removes the internally tracked client count, as we might as well
just use map_size/1, avoiding any risk of having the wrong number.

Change-Id: I89808ab8fb5b05501f14c851521b7e8aacaa07c8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222272
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>